LayoutTests/inspector/extensions/extensions-eval-expected.txt

Tests WebInspector extension API

Started extension.
Running tests...
RUNNING TEST: extension_testEvalConsoleAPI
Evaluate: "function" (exception: undefined)
RUNNING TEST: extension_testEvalDefinesGlobalSymbols
Evaluate: {"func":"undefined","variable":42} (exception: undefined)
RUNNING TEST: extension_testEvalOk
Evaluate: {"str":"foo","num":42} (exception: undefined)
RUNNING TEST: extension_testEvalStatement
Evaluate: 0 (exception: undefined)
RUNNING TEST: extension_testEvalStringifyingLoopFailed
log: Extension server error: Inspector protocol error: Object has too long reference chain(must not be longer than 1000)
Evaluate: undefined (exception: {"code":"E_PROTOCOLERROR","description":"Inspector protocol error: %s","details":["Object has too long reference chain(must not be longer than 1000)"],"isError":true})
RUNNING TEST: extension_testEvalThrows
Evaluate: undefined (exception: {"isException":true,"value":"testExceptionString"})
RUNNING TEST: extension_testEvalUndefined
Evaluate: undefined (exception: undefined)
RUNNING TEST: extension_testWithStringifyOverridden
Evaluate: "OK" (exception: undefined)
All tests done.
